removed unnecessary files for attention and positions as positions will be fixed in fms main and attention is moved to paged_llama
