+++ fms_extras/modules/speculator.py
@@ -1,40 +1,31 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-
 from fms.modules.layernorm import LayerNormParameterized
 
-
 class Speculator(nn.Module):
-    def __init__(self, emb_dim=4096, vocab_size=32000, n_predict=3):
+    def __init__(self, emb_dim=4096, inner_dim=0, vocab_size=32000, n_predict=3):
         super().__init__()
         self.n_predict = n_predict
         self.emb_dim = emb_dim
+        inner_dim = inner_dim if inner_dim!=0 else emb_dim
+        self.inner_dim = inner_dim
         self.vsize = vocab_size
-        self.emb = nn.ModuleList(
-            [nn.Embedding(vocab_size, emb_dim) for _ in range(n_predict)]
-        )
-        self.proj = nn.ModuleList(
-            [nn.Linear(emb_dim * 2, emb_dim, bias=False) for _ in range(n_predict)]
-        )
-        self.head = nn.ModuleList(
-            [nn.Linear(emb_dim, vocab_size, bias=False) for _ in range(n_predict)]
-        )
+        self.emb = nn.ModuleList([nn.Embedding(vocab_size, inner_dim) for _ in range(n_predict)])
+        self.proj = nn.ModuleList([nn.Linear((emb_dim if i==0 else inner_dim), inner_dim, bias=False) for i in range(n_predict)])
+        self.head = nn.ModuleList([nn.Linear(inner_dim, vocab_size, bias=False) for _ in range(n_predict)])
         self.ln = nn.ModuleList(
-            [
-                LayerNormParameterized(
-                    emb_dim, elementwise_shift=True, elementwise_scale=True
-                )
-                for _ in range(n_predict)
-            ]
+            [LayerNormParameterized(inner_dim, elementwise_shift=True, elementwise_scale=True) for _ in range(n_predict)]
         )
+        # Weights ensure that state_0 accounts for 50% of state magnitude by final head in expectation
+        self.state_weight = .5**(.5 / n_predict)
+        self.emb_weight = (1-self.state_weight**2)**.5
         self.a = nn.GELU()
-        self.reset_params()
 
-    def reset_params(self):
+    def reset_parameters(self):
         for m in self.modules():
             if isinstance(m, nn.Embedding) or isinstance(m, nn.Linear):
-                nn.init.trunc_normal_(m.weight, 0, 1 / self.emb_dim**0.5)
+                nn.init.trunc_normal_(m.weight, 0, 1 / self.inner_dim**0.5)
             elif isinstance(m, LayerNormParameterized):
                 m.weight.data.fill_(1)
                 m.bias.data.zero_()
@@ -59,9 +50,8 @@ class Speculator(nn.Module):
         ), f"You must provide a topk number for each head ({self.n_predict} heads, {len(topk)} provided)"
         for i in range(self.n_predict):
             # Project and predict
-            z = self.emb[i](ind)  # b k d
-            z = torch.cat([state, z], dim=2)  # b k 2d
-            state = self.a(self.ln[i](self.proj[i](z)))  # b k d
+            z = self.emb[i](ind).mul(self.emb_weight*(self.inner_dim/2)**.5)  # b k d
+            state = self.a(self.ln[i](self.proj[i](state)*self.state_weight+z))  # b k d
             probs = F.log_softmax(self.head[i](state), dim=2)  # b k v
             probs, preds = probs.topk(topk[i], dim=2)  # b k k'
 
@@ -89,15 +79,13 @@ class Speculator(nn.Module):
         ----
         Since we're assuming all prior tokens are "correct", don't act recursively, just pull from provided inds.
         Produces self.n_predict predicted tokens for each token embedding in state.
-        Inds requires self.n_predict-1 extra tokens on the right to "simulate" recursive behavior for end positions.
+        Inds requires self.n_predict extra tokens on the right to "simulate" recursive behavior for end positions.
         """
         # state: b n d
-        # inds: b n+2 (..., pred token, n+2, n+3)
+        # inds: b n+h (..., pred token, n+2, n+3, n+4)
         out = []
         for i in range(self.n_predict):
-            h_inds = inds[:, i : i + state.size(1)]
-            z = self.emb[i](h_inds)  # b n d
-            z = torch.cat([state, z], dim=2)  # b n 2d
-            state = self.a(self.ln[i](self.proj[i](z)))  # b n d
+            z = self.emb[i](inds[:, i : i + state.size(1)]).mul(self.emb_weight*(self.inner_dim/2)**.5)  # b n d
+            state = self.a(self.ln[i](self.proj[i](state)*self.state_weight+z))  # b n d
             out.append(self.head[i](state))  # b n v
         return torch.stack(out, dim=0)  # h b n v
